requirements.txt
streamlit>=1.34.0
sarvamai>=0.1.0
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
pypdf>=4.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rapidfuzz>=3.9.0
tiktoken>=0.8.0
